feat: prevent mangle name of function

+++ src/lib.rs
@@ -1,4 +1,5 @@
-pub fn add(left: usize, right: usize) -> usize {
+#[no_mangle]
+pub fn add(left: u32, right: u32) -> u32 {
     left + right
 }
